feat: add user-like DTO, notification service interface, and game type command bodies; refactor event service imports and update unit of work methods

+++ src/app/commands/event/delete-event/delete-event.command.ts
@@ -2,6 +2,7 @@ import ICommand from "src/common/app/command.i";
 import { DeleteEventBody } from "./delete-event.body";
 import { IUnitOfWork } from "src/common/domain/unit-of-work.i";
 import { IEventRepository } from "src/domain/event/event.repo.i";
+import { DomainError } from "src/common/error/domain.error";
 
 export class DeleteEventCommand implements ICommand<DeleteEventBody> {
 	constructor(
@@ -11,7 +12,10 @@ export class DeleteEventCommand implements ICommand<DeleteEventBody> {
 	async execute(param: DeleteEventBody): Promise<void> {
 		const eventRepository = this.unitOfWork.getRepository(IEventRepository);
 		const event = await eventRepository.getById(param.eventId);
+		if (!event) {
+			throw new DomainError("Event not found");
+		}
 		await eventRepository.delete(event);
-		await this.unitOfWork.executeTransactions();
+		await this.unitOfWork.commit();
 	}
 }
